Trimmed down tools
