+++ exotica/src/Tools/Conversions.cpp
@@ -1,5 +1,4 @@
 #include <exotica/Tools/Conversions.h>
-#include <exotica/Tools/Exception.h>
 #include <algorithm>
 
 namespace Eigen
@@ -29,5 +28,57 @@ namespace exotica
 
         if(std::find(vec.begin(),vec.end(),key)==vec.end()) return false;
         return true;
+    }
+
+    void vectorExoticaToEigen(const exotica::Vector & exotica,
+        Eigen::VectorXd & eigen)
+    {
+      eigen.resize(exotica.data.size());
+      for (int i = 0; i < exotica.data.size(); i++)
+        eigen(i) = exotica.data[i];
+    }
+
+    void vectorEigenToExotica(Eigen::VectorXd eigen, exotica::Vector & exotica)
+    {
+      exotica.data.resize(eigen.rows());
+      for (int i = 0; i < eigen.rows(); i++)
+        exotica.data[i] = eigen(i);
+    }
+
+    void matrixExoticaToEigen(const exotica::Matrix & exotica,
+        Eigen::MatrixXd & eigen)
+    {
+      if (exotica.col == 0 || exotica.row == 0 || exotica.data.size() == 0)
+      {
+        throw_pretty("Matrix conversion failed, no data in the matrix.");
+      }
+      if (exotica.col * exotica.row != exotica.data.size())
+      {
+        throw_pretty(
+            "Matrix conversion failed, size mismatch."<<exotica.col<<" * "<<exotica.row<<" != "<<exotica.data.size());
+      }
+      eigen.resize(exotica.row, exotica.col);
+      int cnt = 0;
+      for (int r = 0; r < exotica.row; r++)
+        for (int c = 0; c < exotica.col; c++)
+        {
+          eigen(r, c) = exotica.data[cnt];
+          cnt++;
+        }
+    }
+
+    void matrixEigenToExotica(const Eigen::MatrixXd & eigen,
+        exotica::Matrix & exotica)
+    {
+      exotica.row = eigen.rows();
+      exotica.col = eigen.cols();
+      exotica.data.resize(exotica.col * exotica.row);
+      int cnt = 0;
+      for (int r = 0; r < exotica.row; r++)
+        for (int c = 0; c < exotica.col; c++)
+        {
+          exotica.data[cnt] = eigen(r, c);
+          cnt++;
+        }
     }
 }
